# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for managing Fanuc CNC connections and data polling",
        "title": "Fanuc Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/connect": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "If 'id' is provided, checks health of specific connection. If not, lists all connections.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connection"
                ],
                "summary": "Get connections or Check specific connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine ID (optional)",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fanucService.ConnectionResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Connects to a Fanuc machine. Throws 500 if connection takes longer than 5 seconds.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connection"
                ],
                "summary": "Create a new connection",
                "parameters": [
                    {
                        "description": "Connection Data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fanucService.ConnectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fanucService.ConnectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fanucService.ConnectionResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/fanucService.ConnectionResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Connection"
                ],
                "summary": "Delete connection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fanucService.ConnectionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "fanucService.ConnectionRequest": {
            "type": "object",
            "required": [
                "endpoint",
                "series"
            ],
            "properties": {
                "endpoint": {
                    "description": "ip:port",
                    "type": "string"
                },
                "series": {
                    "description": "\"0i\", \"31i\", etc",
                    "type": "string"
                },
                "timeout": {
                    "description": "ms, default 5000",
                    "type": "integer"
                }
            }
        },
        "fanucService.ConnectionResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}